Move internal/unix_noeintr to daemon/internal/unix_noeintr
