+++ daemon/server/router/container/notify_linux.go
@@ -6,7 +6,7 @@ import (
 	"syscall"
 
 	"github.com/containerd/log"
-	"github.com/docker/docker/internal/unix_noeintr"
+	"github.com/docker/docker/daemon/internal/unix_noeintr"
 	"golang.org/x/sys/unix"
 )
