+++ daemon/volume/safepath/k8s_safeopen_linux.go
@@ -23,7 +23,7 @@ import (
 	"strings"
 
 	"github.com/containerd/log"
-	"github.com/docker/docker/internal/unix_noeintr"
+	"github.com/docker/docker/daemon/internal/unix_noeintr"
 	"golang.org/x/sys/unix"
 )
